requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
PyPDF2==3.0.1
python-dotenv==0.19.2
email-validator==1.1.3
Flask-WTF==1.0.0
pytest==7.0.0
pytest-flask==1.2.0
google-generativeai==0.3.1
PyMuPDF==1.25.5
requests==2.28.1
protobuf==3.20.3
google-api-core==2.11.0
google-auth==2.17.3
markupsafe==2.0.1
itsdangerous==2.0.1
jinja2==3.0.3 